requirements.txt
streamlit
langchain
langchain-google-genai
python-dotenv
pypdf
faiss-cpu
-e .
